Rename classycle -> packageCycles
