+++ subprojects/antlr/build.gradle.kts
@@ -35,6 +35,6 @@ dependencies {
     integTestDistributionRuntimeOnly(project(":distributions-full"))
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/api/plugins/antlr/internal/*")
 }
