+++ subprojects/base-services/build.gradle.kts
@@ -49,7 +49,7 @@ dependencies {
     jmh(libs.guava)
 }
 
-classycle {
+packageCycles {
     // Needed for the factory methods in the base class
     excludePatterns.add("org/gradle/util/GradleVersion**")
 }
