+++ subprojects/core/build.gradle.kts
@@ -179,7 +179,7 @@ strictCompile {
     ignoreRawTypes() // raw types used in public API
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/**")
 }
