+++ subprojects/dependency-management/build.gradle.kts
@@ -115,7 +115,7 @@ dependencies {
     crossVersionTestImplementation(libs.jettyWebApp)
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/**")
 }
