+++ subprojects/ear/build.gradle.kts
@@ -38,6 +38,6 @@ strictCompile {
     ignoreRawTypes() // raw types used in public API
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/plugins/ear/internal/*")
 }
