+++ subprojects/kotlin-dsl-provider-plugins/build.gradle.kts
@@ -33,6 +33,6 @@ dependencies {
     testImplementation(libs.mockitoKotlin2)
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/kotlin/dsl/provider/plugins/precompiled/tasks/**")
 }
