+++ subprojects/kotlin-dsl/build.gradle.kts
@@ -117,7 +117,7 @@ dependencies {
     integTestDistributionRuntimeOnly(project(":distributions-basics"))
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/kotlin/dsl/**")
 }
