+++ subprojects/logging/build.gradle.kts
@@ -42,6 +42,6 @@ dependencies {
     integTestDistributionRuntimeOnly(project(":distributions-core"))
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/internal/featurelifecycle/**")
 }
