+++ subprojects/platform-jvm/build.gradle.kts
@@ -49,7 +49,7 @@ strictCompile {
     ignoreDeprecations() // most of this project has been deprecated
 }
 
-classycle {
+packageCycles {
     // Needed for the factory methods in the interface
     excludePatterns.add("org/gradle/jvm/toolchain/JavaLanguageVersion**")
     excludePatterns.add("org/gradle/jvm/toolchain/internal/**")
