+++ subprojects/plugins/build.gradle.kts
@@ -74,7 +74,7 @@ strictCompile {
     ignoreDeprecations() // uses deprecated software model types
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/**")
 }
