+++ subprojects/process-services/build.gradle.kts
@@ -21,6 +21,6 @@ dependencies {
     integTestDistributionRuntimeOnly(project(":distributions-core"))
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/process/internal/**")
 }
