+++ subprojects/signing/build.gradle.kts
@@ -37,7 +37,7 @@ strictCompile {
     ignoreRawTypes() // raw types used in public API
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/plugins/signing/**")
 }
