+++ subprojects/testing-base/build.gradle.kts
@@ -58,7 +58,7 @@ strictCompile {
     ignoreRawTypes() // raw types used in public API (org.gradle.api.tasks.testing.AbstractTestTask)
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/api/internal/tasks/testing/**")
 }
