+++ subprojects/tooling-api/build.gradle.kts
@@ -78,7 +78,7 @@ strictCompile {
     ignoreRawTypes() // raw types used in public API
 }
 
-classycle {
+packageCycles {
     excludePatterns.add("org/gradle/tooling/**")
 }
